refactor: UI 컴포넌트 구조 개선
