+++ src/shared/ui/Textarea/index.tsx
@@ -1,0 +1,22 @@
+import * as React from "react"
+import { forwardRef } from "react"
+
+interface TextareaProps extends React.TextareaHTMLAttributes<HTMLTextAreaElement> {
+  className?: string
+}
+
+/**
+ * @param className - 컴포넌트에 적용할 클래스 이름
+ * @param props - 텍스트 입력 영역에 적용할 속성
+ * @returns 텍스트 입력 영역 컴포넌트
+ */
+export const Textarea = forwardRef<HTMLTextAreaElement, TextareaProps>(({ className, ...props }, ref) => {
+  return (
+    <textarea
+      className={`flex min-h-[150px] w-full rounded-md border border-input bg-white px-3 py-2 text-sm ring-offset-background placeholder:text-muted-foreground focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-2 disabled:cursor-not-allowed disabled:opacity-50 ${className || ""}`}
+      ref={ref}
+      {...props}
+    />
+  )
+})
+Textarea.displayName = "Textarea"
